requirements.dev.numpy.txt
esprima==4.0.1
pytest==7.1.2
pytest-cov==3.0.0
pytest-xdist==2.5.0
python-dotenv>=0.20.0
pytz==2022.1
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinx-toolbox==3.4.0
sphinx-tabs==3.4.1
tox==4.4.6
requests==2.31.0
urllib3==1.26.9
validator-collection==1.5.0
anthropic==0.3.11
dill==0.3.7
openai==0.28.0
IPython>=8.10.0
pyspark>=3.3.0
pandas>=1.3.3
orjson>=3.7.7
